requirements.txt
astroid==1.6.3
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-countries==5.3
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
Pillow==5.1.0
PyJWT==1.6.4
pylint==1.8.4
python-monkey-business==1.0.0
pytz==2018.4
requests==2.18.4
six==1.11.0
transliterate==1.10
urllib3==1.22
uuid==1.30
wrapt==1.10.11
